+++ mrfmsim_plot/__init__.py
@@ -1,4 +1,0 @@
-from mrfmsim_plot.mayaviplots import mayavi_image_plane, mayavi_save
-from mrfmsim_plot.pvplots import pv_imagedata, pv_save, pv_volume
-
-__mrfmsim_plugin__ = ["mayavi_image_plane", "mayavi_save", "pv_imagedata", "pv_save", "pv_volume"]