+++ tests/test_mayaviplot.py
@@ -1,4 +1,4 @@
-from mrfmsim_plot.mayaviplots import mayavi_image_plane, mayavi_save
+from mrfmsim_plot.mayaviplot import mayavi_image_plane
 import numpy as np
 from mayavi import mlab
 import pytest
@@ -21,9 +21,7 @@ class TestMayaviImagePlane:
         """Test mayavi image data."""
 
         engine = mayavi_image_plane(dataset, grid).get_engine()
-        assert np.array_equal(
-            engine.current_scene.children[0].scalar_data, dataset
-        )
+        assert np.array_equal(engine.current_scene.children[0].scalar_data, dataset)
 
     def test_image_plane_widgets(self, dataset, grid, offscreen):
         """Test image_plane widgets start at the right index."""
@@ -37,7 +35,7 @@ class TestMayaviImagePlane:
         assert ipw_y.ipw.slice_index == 10
         assert ipw_z.ipw.plane_orientation == "z_axes"
         assert ipw_z.ipw.slice_index == 10
-    
+
     def test_image_axes(self, dataset, grid, offscreen):
         """Test if the image has the correct extent (or range)."""
         engine = mayavi_image_plane(dataset, grid).get_engine()
@@ -45,24 +43,8 @@ class TestMayaviImagePlane:
         axes = pipeline[4].axes
         assert np.array_equal(axes.ranges, [-45, 55, -40, 60, -15, 5])
 
-    
     def test_image_size(self, dataset, grid, offscreen):
         """Test if the image has the correct size."""
         engine = mayavi_image_plane(dataset, grid, size=(800, 800)).get_engine()
         scene = engine.current_scene
         assert all(scene.scene.get_size() == (800, 800))
-
-def test_mayavi_plot(dataset, grid, tmpdir):
-    """Test the mayavi_plot decorator to see if it saved an image.
-    
-    Here, we only test that it created a file with the correct name.
-    And the file has content.
-    """
-
-    @mayavi_save
-    def plot(dataset, grid):
-        return mayavi_image_plane(dataset, grid)
-
-    plot(dataset, grid, filename=str(tmpdir.join("test.png")))
-    assert tmpdir.join("test.png").check()
-    assert tmpdir.join("test.png").size() > 0